Relocate shared types out of Lambda directory
